chore: rename to get verified fetch
